requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0

SQLAlchemy==2.0.43

pydantic==2.11.7
python-dotenv==1.1.1

google-ads==28.0.0

httpx==0.28.1
pytest==8.4.1